# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cette API permet de gérer les utilisateurs et leurs tâches",
        "title": "API Utilisateurs et Tâches",
        "contact": {
            "name": "Développeur API",
            "email": "mayssa@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/tasks/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extraire les tâches avec de tout les utilisateurs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Extraire les tâches",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Création d'une tâche avec les champs JSON fournis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Créer une tâche",
                "parameters": [
                    {
                        "description": "Les données de tache à créer",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
        "/api/tasks/filtre_date": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Filtrer les taches par date avec limite",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Filtrer les taches par date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date de début",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date de fin",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nombre des jours en arrière",
                        "name": "jours",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
        "/api/tasks/filtrer": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "filterer les tâches par l'ID de l'utilisateur et la status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "filterer les tâches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "L'ID de l'utilisateur",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status du tâche",
                        "name": "completed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
        "/api/tasks/paginated": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extraire les tâches avec pagination, en fonction du page et limit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Extraire les tâches avec pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Les pages",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "La limite des elements",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Tâche introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
        "/api/tasks/rate/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Taux de completion des tâches par l'utilisateur",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Taux de completion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "L'ID de l'utilisateur",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CompletionRate"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extraire une tâche avec son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Extraire une tâche",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la tâche (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Tâche introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mettre à jour les informations d'une tâche avec son ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tâche"
                ],
                "summary": "Mettre à jour une tâche",
                "parameters": [
                    {
                        "type": "string",
                        "description": "L'ID du tâche",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelles données du Tâche",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Tâche introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
...
        "/api/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extraire un utilisatuer avec son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Extraire un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de l'itulisateur (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mettre à jour les informations d'un utilisateur avec son ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Mettre à jour un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "L'ID de l'utilisateur",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nouvelles données utilisateur",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Supprimer un utilisateur par son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Supprimer un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "L'ID de l'utilisateur",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mise à jour partielle de l'utilisateur par son ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Mise à jour partielle de l'utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "L'ID de l'utilisateur",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Les nouvelles données de l'utilisateur",
                        "name": "updateUser",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.UpdateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AppSuccessCRUD"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable",
                        "schema": {
                            "$ref": "#/definitions/utils.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Task": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date_naissance": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "files": {
                    "description": "Foreign key (fichiers)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nom": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tasks": {
                    "description": "Foreign key (taches)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "response.CompletionRate": {
            "type": "object",
            "properties": {
                "completion_rate": {
                    "type": "string"
                }
            }
        },
...
        "response.UpdateUser": {
            "type": "object",
            "properties": {
                "nom": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                }
            }
        },
        "utils.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "utils.AppSuccessCRUD": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Saisir le token JWT comme suit : Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}